split data plugin into smaller http, websql and camera plugins
