+++ src/vue-plugins/camera-cordova/index.js
@@ -1,0 +1,7 @@
+import module from './module';
+
+export default {
+  install(Vue, { store }) {
+    store.registerModule('camera', module);
+  },
+};
