+++ src/vue-plugins/camera-cordova/module.js
